Changed the folder name and service name /domain name
