+++ jazz_test-lambda/Jenkinsfile
@@ -3,10 +3,10 @@ import groovy.json.JsonOutput
 import groovy.transform.Field
 
 // To be replaced as @Field def repo_credential_id = "value" except for config_loader
-@Field def repo_credential_id
-@Field def repo_base
-@Field def repo_core
-@Field def scm_type
+@Field def repo_credential_id = "jenkins1cred"
+@Field def repo_base = "jazz-daily-installer-bitbucket-1723618883.us-east-1.elb.amazonaws.com"
+@Field def repo_core = "slf"
+@Field def scm_type = "bitbucket"
 
 @Field def config_loader
 
@@ -37,7 +37,7 @@ node {
     def build_job = config_loader.API.API_BUILD_URI
     def job_url = JenkinsLocationConfiguration.get().getUrl() + build_job
 
-    sh "curl -X GET -k -v -u \"$UNAME:$PWD\"  \"" + job_url + "&service_name=test-api&domain=test&scm_branch=${env.BRANCH_NAME}\""
+    sh "curl -X GET -k -v -u \"$UNAME:$PWD\"  \"" + job_url + "&service_name=test-lambda&domain=jazz&scm_branch=${env.BRANCH_NAME}\""
     }
   }
 }
