+++ jazz_test-lambda/config/stg-config.json
@@ -1,0 +1,3 @@
+{
+  "AWS_Region": "{conf-region}"
+}
