+++ jazz_test-lambda/deployment-env.yml
@@ -1,0 +1,1 @@
+service_id: {service_id}
