+++ jazz_test-lambda/swagger/swagger.json
@@ -2,16 +2,16 @@
   "swagger": "2.0",
   "info": {
     "version": "2017-01-30T23:30:18Z",
-    "title": "jazz20180529-dev",
+    "title": "{api_deployment_node_title}",
     "description": "API specification for nodejs template service"
   },
-  "host": "dpcfmpwrtl.execute-api.us-east-1.amazonaws.com",
-  "basePath": "/dev/sr8y4mho0e-dev/dev",
+  "host": "{api_host_name}",
+  "basePath": "/{domain}",
   "schemes": [
     "https"
   ],
   "paths": {
-    "/test-lambda": {
+    "/{service_name}": {
       "options": {
         "description": "CORS Support",
         "consumes": [
@@ -214,9 +214,9 @@
         },
         "x-amazon-apigateway-integration": {
           "type": "aws",
-          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:192006145812:function:jazz20180529-dev-test-lambda-sr8y4mho0e-dev/invocations",
+          "uri": "arn:aws:apigateway:{conf-region}:lambda:path/2015-03-31/functions/arn:aws:lambda:{conf-region}:{conf-accId}:function:{lambda_function_name}/invocations",
           "httpMethod": "POST",
-          "credentials": "arn:aws:iam::192006145812:role/jazz20180529_lambda2_basic_execution_1",
+          "credentials": "{conf-role}",
           "requestTemplates": {
             "application/json": "\n#define( $loop )\n  {\n  #foreach($key in $map.keySet())\n      \"$util.escapeJavaScript($key)\":\n        \"$util.escapeJavaScript($map.get($key))\"\n        #if( $foreach.hasNext ) , #end\n  #end\n  }\n#end\n\n{\n  \"body\": $input.json(\"$\"),\n  \"method\": \"$context.httpMethod\",\n  \"principalId\": \"$context.authorizer.principalId\",\n  \"stage\": \"$context.stage\",\n\n  #set( $map = $input.params().header )\n  \"headers\": $loop,\n\n  #set( $map = $input.params().querystring )\n  \"query\": $loop,\n\n  #set( $map = $input.params().path )\n  \"path\": $loop,\n\n  #set( $map = $context.identity )\n  \"identity\": $loop,\n\n  #set( $map = $stageVariables )\n  \"stageVariables\": $loop\n}",
             "application/x-www-form-urlencoded": "\n#define( $body )\n  {\n  #foreach( $token in $input.path('$').split('&') )\n    #set( $keyVal = $token.split('=') )\n    #set( $keyValSize = $keyVal.size() )\n    #if( $keyValSize >= 1 )\n      #set( $key = $util.urlDecode($keyVal[0]) )\n      #if( $keyValSize >= 2 )\n        #set( $val = $util.urlDecode($keyVal[1]) )\n      #else\n        #set( $val = '' )\n      #end\n      \"$key\": \"$val\"#if($foreach.hasNext),#end\n    #end\n  #end\n  }\n#end\n\n#define( $loop )\n  {\n  #foreach($key in $map.keySet())\n      \"$util.escapeJavaScript($key)\":\n        \"$util.escapeJavaScript($map.get($key))\"\n        #if( $foreach.hasNext ) , #end\n  #end\n  }\n#end\n\n{\n  \"body\": $body,\n  \"method\": \"$context.httpMethod\",\n  \"principalId\": \"$context.authorizer.principalId\",\n  \"stage\": \"$context.stage\",\n\n  #set( $map = $input.params().header )\n  \"headers\": $loop,\n\n  #set( $map = $input.params().querystring )\n  \"query\": $loop,\n\n  #set( $map = $input.params().path )\n  \"path\": $loop,\n\n  #set( $map = $context.identity )\n  \"identity\": $loop,\n\n  #set( $map = $stageVariables )\n  \"stageVariables\": $loop\n}"
@@ -386,9 +386,9 @@
           "type": "number",
           "description": "Sample key in the output object"
         },
-        "funcResponse": {
-          "type": "string",
-          "description": "Sample key in the output object"
+        "StatusCode": {
+          "type": "number",
+          "description": "Status of Execution"
         }
       }
     }
