--- test-functions-api/config/stg-config.json
@@ -1,3 +1,0 @@
-{
-  "configKey": "ConfigValueStg"
-}
