--- test-functions-api/deployment-env.yml
@@ -1,1 +1,0 @@
-service_id: 1069f23c-824b-c0de-b73c-58d728cb21ab
